Create trace package
