+++ pkg/client/client_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	. "github.com/keboola/go-client/pkg/client"
+	. "github.com/keboola/go-client/pkg/client/trace"
 	. "github.com/keboola/go-client/pkg/request"
 )
 
@@ -388,8 +389,8 @@ func TestStopRetryOnRequestTimeout(t *testing.T) {
 			WaitTimeStart:       40 * time.Millisecond, // <<<<<<<
 			WaitTimeMax:         40 * time.Millisecond,
 		}).
-		AndTrace(func() *Trace {
-			return &Trace{
+		AndTrace(func() *ClientTrace {
+			return &ClientTrace{
 				HTTPRequestRetry: func(_ int, delay time.Duration) {
 					delays = append(delays, delay)
 				},
