+++ pkg/client/trace/trace_test.go
@@ -1,4 +1,4 @@
-package client_test
+package trace_test
 
 import (
 	"context"
@@ -11,10 +11,10 @@ import (
 
 	"github.com/davecgh/go-spew/spew"
 	"github.com/jarcoal/httpmock"
-	"github.com/keboola/go-utils/pkg/wildcards"
 	"github.com/stretchr/testify/assert"
 
 	. "github.com/keboola/go-client/pkg/client"
+	. "github.com/keboola/go-client/pkg/client/trace"
 	. "github.com/keboola/go-client/pkg/request"
 )
 
@@ -58,8 +58,8 @@ func TestTrace(t *testing.T) {
 			WaitTimeStart: 1 * time.Microsecond,
 			WaitTimeMax:   20 * time.Microsecond,
 		}).
-		AndTrace(func() *Trace {
-			return &Trace{
+		AndTrace(func() *ClientTrace {
+			return &ClientTrace{
 				GotRequest: func(ctx context.Context, request HTTPRequest) context.Context {
 					logs.WriteString(fmt.Sprintf("GotRequest        %s %s\n", request.Method(), request.URL()))
 					return ctx
@@ -123,8 +123,8 @@ func TestTrace_Multiple(t *testing.T) {
 	c := New().
 		WithTransport(transport).
 		WithRetry(TestingRetry()).
-		AndTrace(func() *Trace {
-			return &Trace{
+		AndTrace(func() *ClientTrace {
+			return &ClientTrace{
 				GotRequest: func(ctx context.Context, request HTTPRequest) context.Context {
 					logs.WriteString(fmt.Sprintf("1: GotRequest        %s %s\n", request.Method(), request.URL()))
 					return ctx
@@ -143,8 +143,8 @@ func TestTrace_Multiple(t *testing.T) {
 				},
 			}
 		}).
-		AndTrace(func() *Trace {
-			return &Trace{
+		AndTrace(func() *ClientTrace {
+			return &ClientTrace{
 				GotRequest: func(ctx context.Context, request HTTPRequest) context.Context {
 					logs.WriteString(fmt.Sprintf("2: GotRequest        %s %s\n", request.Method(), request.URL()))
 					return ctx
@@ -157,8 +157,8 @@ func TestTrace_Multiple(t *testing.T) {
 				},
 			}
 		}).
-		AndTrace(func() *Trace {
-			return &Trace{
+		AndTrace(func() *ClientTrace {
+			return &ClientTrace{
 				RequestProcessed: func(result any, err error) {
 					s := spew.NewDefaultConfig()
 					s.DisablePointerAddresses = true
@@ -194,123 +194,4 @@ func TestTrace_Multiple(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, "OK", *result.(*string))
 	assert.Equal(t, strings.TrimLeft(expected, "\n"), logs.String())
-}
-
-func TestDumpTracer(t *testing.T) {
-	t.Parallel()
-
-	// Mocked response
-	transport := httpmock.NewMockTransport()
-	transport.RegisterResponder("GET", `https://example.com`, httpmock.ResponderFromMultipleResponses([]*http.Response{
-		{StatusCode: http.StatusLocked},
-		{StatusCode: http.StatusTooManyRequests},
-		{StatusCode: http.StatusOK, Body: io.NopCloser(strings.NewReader("OK"))},
-	}))
-
-	// Logs for trace testing
-	var logs strings.Builder
-
-	// Create client
-	ctx := context.Background()
-	c := New().
-		WithTransport(transport).
-		WithRetry(TestingRetry()).
-		AndTrace(DumpTracer(&logs))
-
-	// Expected trace
-	expected := `
->>>>>> HTTP DUMP
-GET / HTTP/1.1
-Host: example.com
-User-Agent: keboola-go-client
-Accept-Encoding: gzip, br
-------
-HTTP/0.0 423 Locked
-Content-Length: 0
-<<<<<< HTTP DUMP END
-
->>>>>> HTTP RETRY | ATTEMPT: 1 | DELAY: 1ms |  GET / 423 | ERROR: <nil>
-
->>>>>> HTTP DUMP
-GET / HTTP/1.1
-Host: example.com
-User-Agent: keboola-go-client
-Accept-Encoding: gzip, br
-------
-HTTP/0.0 429 Too Many Requests
-Content-Length: 0
-<<<<<< HTTP DUMP END
-
->>>>>> HTTP RETRY | ATTEMPT: 2 | DELAY: 1ms |  GET / 429 | ERROR: <nil>
-
->>>>>> HTTP DUMP
-GET / HTTP/1.1
-Host: example.com
-User-Agent: keboola-go-client
-Accept-Encoding: gzip, br
-------
-HTTP/0.0 200 OK
-Content-Length: 0
-------
-OK
-<<<<<< HTTP DUMP END
-
->>>>>> HTTP REQUEST PROCESSED |  GET / 200 | ERROR: <nil> | HEADERS AT: %s | DONE AT: %s
-`
-
-	// Test
-	str := ""
-	_, result, err := NewHTTPRequest(c).WithGet("https://example.com").WithResult(&str).Send(ctx)
-	assert.NoError(t, err)
-	assert.Equal(t, "OK", *result.(*string))
-	wildcards.Assert(t, strings.TrimLeft(expected, "\n"), logs.String())
-}
-
-func TestLogTracer(t *testing.T) {
-	t.Parallel()
-
-	// Mocked response
-	transport := httpmock.NewMockTransport()
-	transport.RegisterResponder("GET", `https://example.com`, httpmock.ResponderFromMultipleResponses([]*http.Response{
-		{StatusCode: http.StatusLocked},
-		{StatusCode: http.StatusTooManyRequests},
-		{StatusCode: http.StatusOK, Body: io.NopCloser(strings.NewReader("OK1"))},
-		{StatusCode: http.StatusOK, Body: io.NopCloser(strings.NewReader("OK2"))},
-	}))
-
-	// Logs for trace testing
-	var logs strings.Builder
-
-	// Create client
-	ctx := context.Background()
-	c := New().
-		WithTransport(transport).
-		WithRetry(TestingRetry()).
-		AndTrace(LogTracer(&logs))
-
-	// Expected trace
-	expected := `
-HTTP_REQUEST[0001] START GET "https://example.com"
-HTTP_REQUEST[0001] DONE  GET "https://example.com" | 423 | %s
-HTTP_REQUEST[0001] RETRY GET "https://example.com" | 1x | 1ms
-HTTP_REQUEST[0001] START GET "https://example.com"
-HTTP_REQUEST[0001] DONE  GET "https://example.com" | 429 | %s
-HTTP_REQUEST[0001] RETRY GET "https://example.com" | 2x | 1ms
-HTTP_REQUEST[0001] START GET "https://example.com"
-HTTP_REQUEST[0001] DONE  GET "https://example.com" | 200 | %s
-HTTP_REQUEST[0001] BODY  GET "https://example.com" | %s
-HTTP_REQUEST[0002] START GET "https://example.com"
-HTTP_REQUEST[0002] DONE  GET "https://example.com" | 200 | %s
-HTTP_REQUEST[0002] BODY  GET "https://example.com" | %s
-`
-
-	// Test
-	str := ""
-	_, result, err := NewHTTPRequest(c).WithGet("https://example.com").WithResult(&str).Send(ctx)
-	assert.NoError(t, err)
-	assert.Equal(t, "OK1", *result.(*string))
-	_, result, err = NewHTTPRequest(c).WithGet("https://example.com").WithResult(&str).Send(ctx)
-	assert.NoError(t, err)
-	assert.Equal(t, "OK2", *result.(*string))
-	wildcards.Assert(t, strings.TrimLeft(expected, "\n"), logs.String())
 }
